requirements.txt
blessings==1.7
PyYAML==5.3.1
six==1.14.0
terminal==0.4.0
